Refactor extract heading level function
