+++ apps/web/components/Organization/Slice/AccordionSlice/AccordionSlice.tsx
@@ -1,5 +1,7 @@
 import React from 'react'
 import { useRouter } from 'next/router'
+
+import { SliceType } from '@island.is/island-ui/contentful'
 import {
   Accordion,
   AccordionCard,
@@ -15,24 +17,8 @@ import {
   AccordionSlice as AccordionSliceSchema,
   Html,
 } from '@island.is/web/graphql/schema'
-import { SliceType } from '@island.is/island-ui/contentful'
+import { extractHeadingLevels } from '@island.is/web/utils/navigation'
 import { webRichText } from '@island.is/web/utils/richText'
-
-const headingLevels = ['h2', 'h3', 'h4', 'h5'] as const
-type HeadingType = typeof headingLevels[number]
-
-export const extractHeadingLevels = (slice: AccordionSliceSchema) => {
-  let titleHeading: HeadingType = 'h2'
-  let childHeading: HeadingType = 'h3'
-
-  if (headingLevels.includes(slice.titleHeadingLevel as HeadingType)) {
-    titleHeading = slice.titleHeadingLevel as HeadingType
-    childHeading = `h${Number(titleHeading[1]) + 1}` as HeadingType
-  }
-
-  return { titleHeading, childHeading }
-}
-
 interface SliceProps {
   slice: AccordionSliceSchema
 }
