+++ apps/web/components/connected/AdministrationOfOccupationalSafetyAndHealthCourses/AdministrationOfOccupationalSafetyAndHealthCourses.tsx
@@ -12,6 +12,7 @@ import {
 import { ConnectedComponent, Query } from '@island.is/web/graphql/schema'
 import { useNamespace } from '@island.is/web/hooks'
 import { useDateUtils } from '@island.is/web/i18n/useDateUtils'
+import { extractHeadingLevels } from '@island.is/web/utils/navigation'
 
 import { GET_ADMINISTRATION_OF_SAFETY_AND_HEALTH_COURSES_QUERY } from './queries'
 import { getCurrencyString, parseDateString } from './utils'
@@ -22,21 +23,19 @@ const normalizesAndMatch = (value1: string, value2: string) => {
 
 interface AdministrationOfOccupationalSafetyAndHealthCoursesProps {
   slice: ConnectedComponent
-  title: string | null
-  headingLevel: 'h2' | 'h3' | 'h4'
-  hasBorderAbove: boolean
 }
 
 type ListState = 'loading' | 'loaded' | 'error'
 
 const AdministrationOfOccupationalSafetyAndHealthCourses = ({
   slice,
-  title = slice.json?.title ?? null,
-  headingLevel = slice.configJson?.headingLevel ?? 'h2',
-  hasBorderAbove = slice.configJson?.hasBorderAbove ?? true,
 }: AdministrationOfOccupationalSafetyAndHealthCoursesProps) => {
   const n = useNamespace(slice.json ?? {})
   const { format } = useDateUtils()
+
+  const title = slice.json?.title ?? null
+  const hasBorderAbove = slice.configJson?.hasBorderAbove ?? false
+  const { titleHeading } = extractHeadingLevels(slice.configJson ?? {})
 
   const [listState, setListState] = useState<ListState>('loading')
   const [courses, setCourses] = useState<
@@ -90,7 +89,7 @@ const AdministrationOfOccupationalSafetyAndHealthCourses = ({
       )}
 
       {title && (
-        <Text variant={headingLevel} as={headingLevel}>
+        <Text variant="h2" as={titleHeading}>
           {title}
         </Text>
       )}
