+++ apps/web/utils/navigation.ts
@@ -2,6 +2,8 @@ import { AllSlicesFragment as Slice } from '@island.is/web/graphql/schema'
 import { Document, BLOCKS, Block, Text } from '@contentful/rich-text-types'
 import slugify from '@sindresorhus/slugify'
 
+const headingLevels = ['h2', 'h3', 'h4', 'h5'] as const
+type HeadingType = typeof headingLevels[number]
 interface NavLink {
   id: string
   text: string
@@ -71,4 +73,18 @@ const sliceToNavLinks = (slice: Slice, htmlTags: BLOCKS[]): NavLink[] => {
   }
 
   return []
+}
+
+export const extractHeadingLevels = (slice: {
+  titleHeadingLevel?: string | null
+}) => {
+  let titleHeading: HeadingType = 'h2'
+  let childHeading: HeadingType = 'h3'
+
+  if (headingLevels.includes(slice.titleHeadingLevel as HeadingType)) {
+    titleHeading = slice.titleHeadingLevel as HeadingType
+    childHeading = `h${Number(titleHeading[1]) + 1}` as HeadingType
+  }
+
+  return { titleHeading, childHeading }
 }
